Move tests to the dedicated package
