+++ selenium/che-selenium-test/src/test/java/org/eclipse/che/selenium/site/ocpoauth/LoginNewUserWithOpenShiftOAuthTest.java
@@ -8,20 +8,22 @@
  * Contributors:
  *   Red Hat, Inc. - initial API and implementation
  */
-package org.eclipse.che.selenium.site;
+package org.eclipse.che.selenium.site.ocpoauth;
 
-import static java.lang.String.format;
 import static org.eclipse.che.selenium.pageobject.dashboard.NewWorkspace.Stack.JAVA;
-import static org.testng.Assert.assertEquals;
 
 import com.google.inject.Inject;
-import org.eclipse.che.api.core.model.workspace.Workspace;
+import org.eclipse.che.api.core.BadRequestException;
+import org.eclipse.che.api.core.ConflictException;
+import org.eclipse.che.api.core.NotFoundException;
+import org.eclipse.che.api.core.ServerException;
+import org.eclipse.che.api.core.model.user.User;
 import org.eclipse.che.commons.lang.NameGenerator;
 import org.eclipse.che.selenium.core.SeleniumWebDriver;
 import org.eclipse.che.selenium.core.TestGroup;
-import org.eclipse.che.selenium.core.client.TestWorkspaceServiceClient;
+import org.eclipse.che.selenium.core.client.TestUserServiceClient;
 import org.eclipse.che.selenium.core.provider.TestDashboardUrlProvider;
-import org.eclipse.che.selenium.core.user.DefaultTestUser;
+import org.eclipse.che.selenium.core.user.TestUser;
 import org.eclipse.che.selenium.core.webdriver.SeleniumWebDriverHelper;
 import org.eclipse.che.selenium.pageobject.Ide;
 import org.eclipse.che.selenium.pageobject.ToastLoader;
@@ -37,8 +39,8 @@ import org.testng.annotations.AfterClass;
 import org.testng.annotations.Test;
 
 /**
- * This test checks if existed user of Eclipse Che Multiuser deployed on OCP can log into Eclipse
- * Che by using OCP OAuth server identity provider and work with workspace as own OCP resource.<br>
+ * This test checks user can be registered in Eclipse Che Multiuser deployed on OCP with OCP OAuth
+ * server identity provider and work with workspace as own OCP resource.<br>
  * <br>
  * <b>Test environment</b>: <br>
  * Eclipse Che Multiuser deployed on OCP with support of OpenShift OAuth server.<br>
@@ -46,19 +48,16 @@ import org.testng.annotations.Test;
  * {@code --setup-ocp-oauth} parameter to setup OCP OAuth identity provider.<br>
  * <br>
  * <b>Test case</b>:<br>
- * - register new Eclipse Che user;<br>
  * - go to login page of Eclipse Che;<br>
  * - click on button to login with OpenShift OAuth;<br>
- * - login to OCP with registered earlier Eclipse Che test user credentials;<br>
+ * - login to OCP with new Eclipse Che test user credentials;<br>
  * - authorize ocp-client to access OpenShift account;<br>
  * - fill first broker profile page;<br>
- * - add OCP user to existed Che user account;<br>
- * - login into Eclipse Che again;<br>
  * - create and open workspace of java type;<br>
  * - switch to the Eclipse Che IDE and wait until workspace is ready to use;<br>
- * - go to OCP and check there if there is a project with name equals to test workspace id;<br>
+ * - go to OCP and check there if there is a project with name starts from "workspace";<br>
  * - remove test workspace from Eclipse Che Dashboard;<br>
- * - go to OCP and check there if there is no project with name equals to test workspace id.<br>
+ * - go to OCP and check there if there is no project with name starts from "workspace".<br>
  * <br>
  * <a href="https://github.com/eclipse/che/issues/8178">Feature reference.</a> <br>
  * <br>
@@ -66,39 +65,40 @@ import org.testng.annotations.Test;
  * @author Dmytro Nochevnov
  */
 @Test(groups = {TestGroup.OPENSHIFT, TestGroup.MULTIUSER})
-public class LoginExistedUserWithOpenShiftOAuthTest {
+public class LoginNewUserWithOpenShiftOAuthTest {
 
   private static final String WORKSPACE_NAME = NameGenerator.generate("workspace", 4);
+  private static final String WORKSPACE_ID_PREFIX = "workspace";
 
-  public static final String LOGIN_TO_CHE_WITH_OPENSHIFT_OAUTH_MESSAGE_TEMPLATE =
-      "Authenticate as %s to link your account with openshift-v3";
-
-  public static final String USER_ALREADY_EXISTS_ERROR_MESSAGE_TEMPLATE =
-      "User with email %s already exists. How do you want to continue?";
+  private static final TestUser NEW_TEST_USER = getTestUser();
 
   @Inject private CheLoginPage cheLoginPage;
   @Inject private OpenShiftLoginPage openShiftLoginPage;
   @Inject private FirstBrokerProfilePage firstBrokerProfilePage;
   @Inject private AuthorizeOpenShiftAccessPage authorizeOpenShiftAccessPage;
-  @Inject private DefaultTestUser defaultTestUser;
   @Inject private Dashboard dashboard;
   @Inject private Workspaces workspaces;
   @Inject private NewWorkspace newWorkspace;
-  @Inject private TestWorkspaceServiceClient defaultUserWorkspaceServiceClient;
   @Inject private ToastLoader toastLoader;
   @Inject private Ide ide;
   @Inject private SeleniumWebDriverHelper seleniumWebDriverHelper;
+  @Inject private TestUserServiceClient testUserServiceClient;
   @Inject private OpenShiftProjectCatalogPage openShiftProjectCatalogPage;
   @Inject private SeleniumWebDriver seleniumWebDriver;
   @Inject private TestDashboardUrlProvider testDashboardUrlProvider;
 
   @AfterClass
-  private void removeTestWorkspace() throws Exception {
-    defaultUserWorkspaceServiceClient.delete(WORKSPACE_NAME, defaultTestUser.getName());
+  private void removeTestUser() throws ServerException, ConflictException, BadRequestException {
+    try {
+      User newTestUserDto = testUserServiceClient.findByName(NEW_TEST_USER.getName());
+      testUserServiceClient.remove(newTestUserDto.getId());
+    } catch (NotFoundException e) {
+      // ignore if test user don't exist
+    }
   }
 
   @Test
-  public void checkExistedCheUserOcpProjectCreationAndRemoval() throws Exception {
+  public void checkNewCheUserOcpProjectCreationAndRemoval() {
     // go to login page of Eclipse Che
     // (we can't use dashboard.open() here to login with OAuth)
     seleniumWebDriver.navigate().to(testDashboardUrlProvider.get());
@@ -106,28 +106,15 @@ public class LoginExistedUserWithOpenShiftOAuthTest {
     // click on button to login with OpenShift OAuth
     cheLoginPage.loginWithOpenShiftOAuth();
 
-    // login to OCP from login page with default test user credentials
-    openShiftLoginPage.login(defaultTestUser.getName(), defaultTestUser.getPassword());
+    // login to OCP from login page with new test user credentials
+    openShiftLoginPage.login(NEW_TEST_USER.getName(), NEW_TEST_USER.getPassword());
 
     // authorize ocp-client to access OpenShift account
     authorizeOpenShiftAccessPage.waitOnOpen();
     authorizeOpenShiftAccessPage.allowPermissions();
 
     // fill first broker profile page
-    firstBrokerProfilePage.submit(defaultTestUser);
-
-    // add OCP user to existed Eclipse Che user account
-    String expectedError =
-        format(USER_ALREADY_EXISTS_ERROR_MESSAGE_TEMPLATE, defaultTestUser.getEmail());
-
-    assertEquals(firstBrokerProfilePage.getErrorAlert(), expectedError);
-    firstBrokerProfilePage.addToExistingAccount();
-
-    // login into Eclipse Che again
-    String expectedInfo =
-        format(LOGIN_TO_CHE_WITH_OPENSHIFT_OAUTH_MESSAGE_TEMPLATE, defaultTestUser.getName());
-    assertEquals(cheLoginPage.getInfoAlert(), expectedInfo);
-    cheLoginPage.loginWithPredefinedUsername(defaultTestUser.getPassword());
+    firstBrokerProfilePage.submit(NEW_TEST_USER);
 
     // create and open workspace of java type
     dashboard.selectWorkspacesItemOnDashboard();
@@ -143,12 +130,10 @@ public class LoginExistedUserWithOpenShiftOAuthTest {
     toastLoader.waitToastLoaderAndClickStartButton();
     ide.waitOpenedWorkspaceIsReadyToUse();
 
-    // go to OCP and check if there is a project with name equals to test workspace id
+    // go to OCP and check if there is a project with name starts from "workspace"
     openShiftProjectCatalogPage.open();
-    openShiftLoginPage.login(defaultTestUser.getName(), defaultTestUser.getPassword());
-    Workspace testWorkspace =
-        defaultUserWorkspaceServiceClient.getByName(WORKSPACE_NAME, defaultTestUser.getName());
-    openShiftProjectCatalogPage.waitProject(testWorkspace.getId());
+    openShiftLoginPage.login(NEW_TEST_USER.getName(), NEW_TEST_USER.getPassword());
+    openShiftProjectCatalogPage.waitProject(WORKSPACE_ID_PREFIX);
 
     // remove test workspace from Eclipse Che Dashboard
     seleniumWebDriver.navigate().to(testDashboardUrlProvider.get());
@@ -158,8 +143,50 @@ public class LoginExistedUserWithOpenShiftOAuthTest {
     workspaces.clickOnDeleteButtonInDialogWindow();
     workspaces.waitWorkspaceIsNotPresent(WORKSPACE_NAME);
 
-    // go to OCP and check if there is no project with name equals to test workspace id
+    // go to OCP and check if there is no project with name starts from "workspace"
     openShiftProjectCatalogPage.open();
-    openShiftProjectCatalogPage.waitProjectAbsence(testWorkspace.getId());
+    openShiftProjectCatalogPage.waitProjectAbsence(WORKSPACE_ID_PREFIX);
+  }
+
+  private static TestUser getTestUser() {
+    return new TestUser() {
+      private final long currentTimeInMillisec = System.currentTimeMillis();
+      private final String name = "user" + currentTimeInMillisec;
+      private final String email = name + "@1.com";
+      private final String password = String.valueOf(currentTimeInMillisec);
+
+      @Override
+      public String getEmail() {
+        return email;
+      }
+
+      @Override
+      public String getPassword() {
+        return password;
+      }
+
+      @Override
+      public String obtainAuthToken() {
+        return null;
+      }
+
+      @Override
+      public String getOfflineToken() {
+        return null;
+      }
+
+      @Override
+      public String getName() {
+        return name;
+      }
+
+      @Override
+      public String getId() {
+        return null;
+      }
+
+      @Override
+      public void delete() {}
+    };
   }
 }
